1. Added non-working trackball code.

+++ Vector.hs
@@ -13,25 +13,27 @@ module Vector
          cross
          ) where
          
-data Vec3 v = MkVec3 Double Double Double deriving Show
--- getx::Vec3 -> Double
-getx::Vec3 Double -> Double
+import Graphics.UI.GLUT
+
+data Vec3 v = MkVec3 GLfloat GLfloat GLfloat deriving Show
+-- getx::Vec3 -> GLfloat
+getx::Vec3 GLfloat -> GLfloat
 getx (MkVec3 x y z) = x
 
-gety::Vec3 Double -> Double
+gety::Vec3 GLfloat -> GLfloat
 gety (MkVec3 x y z) = y
 
-getz::Vec3 Double -> Double
+getz::Vec3 GLfloat -> GLfloat
 getz (MkVec3 x y z) = z
 
-vabs::Vec3 Double -> Double
+vabs::Vec3 GLfloat -> GLfloat
 vabs (MkVec3 x y z) = sqrt ((x*x) + (y*y) + (z*z))
 
-(./.)::Vec3 Double -> Double -> Vec3 Double
+(./.)::Vec3 GLfloat -> GLfloat -> Vec3 GLfloat
 (./.) a m = MkVec3 ((getx a)/m) ((gety a)/m) ((getz a)/m)
 
 -- vector add
-(.+.)::Vec3 Double -> Vec3 Double -> Vec3 Double
+(.+.)::Vec3 GLfloat -> Vec3 GLfloat -> Vec3 GLfloat
 (.+.) a0 a1 = MkVec3 (x0 + x1) (y0 + y1) (z0 + z1) 
               where
                 x0 = getx a0
@@ -41,7 +43,7 @@ vabs (MkVec3 x y z) = sqrt ((x*x) + (y*y) + (z*z))
                 y1 = gety a1
                 z1 = getz a1
 
-(.-.)::Vec3 Double -> Vec3 Double -> Vec3 Double
+(.-.)::Vec3 GLfloat -> Vec3 GLfloat -> Vec3 GLfloat
 (.-.) a0 a1 = MkVec3 (x0 - x1) (y0 - y1) (z0 - z1) -- vector subtraction
               where
                 x0 = getx a0
@@ -51,7 +53,7 @@ vabs (MkVec3 x y z) = sqrt ((x*x) + (y*y) + (z*z))
                 y1 = gety a1
                 z1 = getz a1
 
-dot::Vec3 Double -> Vec3 Double -> Double
+dot::Vec3 GLfloat -> Vec3 GLfloat -> GLfloat
 dot a0 a1 = (x0 * x1) + (y0 * y1) + (z0 * z1) -- dot product
               where
                 x0 = getx a0
@@ -61,14 +63,14 @@ dot a0 a1 = (x0 * x1) + (y0 * y1) + (z0 * z1) -- dot product
                 y1 = gety a1
                 z1 = getz a1
 
-(.*.)::Vec3 Double -> Double -> Vec3 Double 
+(.*.)::Vec3 GLfloat -> GLfloat -> Vec3 GLfloat 
 (.*.) a0 m = MkVec3 (x0*m) (y0*m) (z0*m)
               where
                 x0 = getx a0
                 y0 = gety a0
                 z0 = getz a0
 
-cross::Vec3 Double -> Vec3 Double -> Vec3 Double
+cross::Vec3 GLfloat -> Vec3 GLfloat -> Vec3 GLfloat
 cross a b = MkVec3 (a2 * b3 - a3 * b2) (-(a1 * b3 - a3 * b1)) (a1 * b2 - a2 * b1)
              where
                a1 = getx a
